% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fit.R
\name{computeAUC}
\alias{computeAUC}
\title{Calculate AUC or AAC}
\usage{
computeAUC(dr_data, dose_col = 1, resp_col = 2, aac = FALSE)
}
\arguments{
\item{dr_data}{Dose-response dataframe. If columns are not labeled "DOSE" and "RESPONSE," indicate the indices of the relevant columns.}

\item{dose_col}{Index of the dose/concentration column.}

\item{resp_col}{Index of the response column, percentage viability recommended.}

\item{aac}{Logical. Whether or not to return the Area Above the Dose-Response Curve (AAC).}
}
\value{
A numeric value representing the AUC or AAC.
}
\description{
Calculate the Area Under the Curve (AUC) or the Area Above the Curve (AAC) for a dose-response curve.
}
\examples{
\dontrun{
  data <- data.frame(
    DOSE = c(0, 1, 2, 3, 4),
    RESPONSE = c(100, 90, 80, 70, 60)
  )
  computeAUC(data, dose_col = 1, resp_col = 2)
}

}
