% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_processing_pipeline.R
\name{curve_process_pipeline}
\alias{curve_process_pipeline}
\title{Parallel processing application of curve assessment function}
\usage{
curve_process_pipeline(
  cell_line_col = NULL,
  drug_name_col = NULL,
  dr_dataset,
  n_cores = 4,
  output_fname = NULL
)
}
\arguments{
\item{cell_line_col}{Index for cell line column (if not named "CELL_LINE_NAME").}

\item{drug_name_col}{Index for drug name column (if not named "DRUG_NAME").}

\item{dr_dataset}{Data frame with dose response data in long format (Column for
cell line id, drug id, dose and % viability).}

\item{n_cores}{Number of CPU's to use.}

\item{output_fname}{File name for csv file output. If NULL, file will not be
saved in function.}
}
\value{
Dataframe with curve metrics and DSS scores. Each row pertaining to a
 single dose cuve (identified by cell line / drug id column in output).
}
\description{
Wrapper function for large scale implementation of model fitting and
 curve assessment.
}
