% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fit.R
\name{fit_drc_model}
\alias{fit_drc_model}
\title{Fit dose response curve}
\usage{
fit_drc_model(dr_data, dose_col = 1, resp_col = 2)
}
\arguments{
\item{dr_data}{Dose response dataframe. If columns are not labelled "DOSE" and
"RESPONSE", indicate the indices of the relevant columns.}

\item{dose_col}{Index of dose/concentration column.}

\item{resp_col}{Index of response column. % Viability recommended.}
}
\value{
A 4-parameter drc model if convergence is achieved. Otherwise
 returns NULL.
}
\description{
Function to fit 4 parameter dose response model.
}
