% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dose_response_fit.R
\name{plot_dr_curve}
\alias{plot_dr_curve}
\title{Plot dose response curve}
\usage{
plot_dr_curve(model, dr_data, dose_col = NULL, resp_col = NULL, ic50 = NULL)
}
\arguments{
\item{model}{4-parameter drc model.}

\item{dr_data}{Dose response dataframe. If columns are not labelled "DOSE" and
"RESPONSE", indicate the indices of the relevant columns.}

\item{dose_col}{Index of dose/concentration column.}

\item{resp_col}{Index of response column. % Viability recommended.}

\item{ic50}{IC50 value for highlight on plot.}
}
\value{
Plot of dose response relationship.
}
\description{
Plot dose response curve fit using fit_drc_model function.
}
