% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_random_curves.R
\name{plot_random_drs}
\alias{plot_random_drs}
\title{Plot random dose response curves}
\usage{
plot_random_drs(
  cell_line_col = NULL,
  drug_name_col = NULL,
  dose_col = NULL,
  resp_col = NULL,
  sample_size = 25,
  save_pdf = TRUE,
  report_title = "Random_curves",
  dr_dataset
)
}
\arguments{
\item{cell_line_col}{Index for cell line column (if not named "CELL_LINE_NAME").}

\item{drug_name_col}{Index for drug name column (if not named "DRUG_NAME").}

\item{dose_col}{Index of dose/concentration column.}

\item{resp_col}{Index of response column.}

\item{sample_size}{Number of dose response curves to plot. Note this does not
account for lack of convergence.}

\item{save_pdf}{Logical value, determines whether to save output pdf.}

\item{report_title}{String to be used as pdf file name.}

\item{dr_dataset}{Data frame with dose response data in long format (Column for
cell line id, drug id, dose and percentage viability).}
}
\description{
Plot dose response curves from large dataset using random selection. Intended
 for checking large datasets with many dose response relationships too large
 to plot all possible curves. For example, if user wants to perform visual
 checks on model fits, this function could be used to generate a sample of
 dose response graphs construction.
 Note to self, in the future include option for different model types.
}
