% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_assessement.R
\name{run_curve_assessment}
\alias{run_curve_assessment}
\title{Curve assessment - QC metrics and DSS calculation}
\usage{
run_curve_assessment(model, dr_data, dose_col = NULL, resp_col = NULL)
}
\arguments{
\item{model}{4-parameter drc model.}

\item{dr_data}{Dose response dataframe. If columns are not labelled "DOSE" and
"RESPONSE", indicate the indices of the relevant columns.}

\item{dose_col}{Index of dose/concentration column.}

\item{resp_col}{Index of response column. % Viability recommended.}
}
\value{
Dataframe with single row.
}
\description{
Extract curve parameter estimates along with QC metrics including R2, RMSE
 and parameter confidence intervals extracted from drc model parameters. This
 includes EC_50 (half of maximum inhibition - also known as relative IC50).
 Additionally, Drug Sensitivity Scores are calculated (DSS1, DSS2 & DSS3). For this,
 ensure data is % of viability (100% = DMSO control) on a scale of 0 - 100,
 and that concentration is in uM and not in log10 scale as transformation is
 applied within function.
}
