% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_processing_pipeline.R
\name{curve_process_pipeline}
\alias{curve_process_pipeline}
\title{Parallel processing application of curve assessment function}
\usage{
curve_process_pipeline(
  cell_line_col = NULL,
  drug_name_col = NULL,
  dose_col = NULL,
  resp_col = NULL,
  dr_dataset,
  n_cores = 4,
  output_fname = NULL
)
}
\arguments{
\item{cell_line_col}{Index for the cell line column (if not named "CELL_LINE_NAME").}

\item{drug_name_col}{Index for the drug name column (if not named "DRUG_NAME").}

\item{dose_col}{Index of the dose/concentration column.}

\item{resp_col}{Index of the response column.}

\item{dr_dataset}{Data frame with dose-response data in long format (Columns for
cell line id, drug id, dose, and percentage viability).}

\item{n_cores}{Number of CPUs to use.}

\item{output_fname}{File name for CSV file output. If NULL, the file will not be
saved by the function.}
}
\value{
Data frame with curve metrics and DSS scores. Each row pertains to a
  single dose curve (identified by cell line/drug id columns in the output).
}
\description{
Wrapper function for large-scale implementation of model fitting and
curve assessment.
}
\examples{
\dontrun{
curve_process_pipeline(
  cell_line_col = 1,
  drug_name_col = 2,
  dose_col = 3,
  resp_col = 4,
  dr_dataset = your_data_frame,
  n_cores = 4,
  output_fname = "output.csv"
)
}

}
